+++ CHANGELOG.md
@@ -4,6 +4,9 @@
   * The context menu now indicates when a backup is locked.
     ([Contributed by darklinkpower](https://github.com/mtkennerly/ludusavi-playnite/pull/96))
   * The context menu now lets you open Ludusavi to add/edit a game's custom entry.
+    This requires Ludusavi v0.30.0 or newer.
+  * The context menu now lets you (un)lock a backup and edit its comment.
+    This requires Ludusavi v0.30.0 or newer.
   * Updated translations, including a new translation for Turkish.
     (Thanks to contributors on the [Crowdin project](https://crowdin.com/project/ludusavi-playnite))